refactor: Extract runReportGeneration to report.go
main.go: 1014 → 724 lines
Final extraction - CLI refactor complete: 1682 → 724 lines (57% reduction)
